+++ js/const.js
@@ -1,0 +1,29 @@
+export const NUMBER_TEMPLATES = 25;
+export const MESSAGE_COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const NAME_COMMENTS = [
+  'Артем',
+  'Сергей',
+  'Арсений',
+  'Генадий',
+  'Инакентий',
+  'Юсуф',
+  'Таштимир',
+];
+
+export const DESCRIPTION = [
+  'Отдыхаю',
+  'Загораю',
+  'Подаю',
+  'Уничтожаю',
+  'Разжигаю',
+  'Поминаю',
+  'Оплакиваю',
+];
